% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientClass.R
\name{age,Patient,Date-method}
\alias{age,Patient,Date-method}
\title{Title}
\usage{
\S4method{age}{Patient,Date}(object, date, units = "years")
}
\arguments{
\item{object}{.}

\item{date}{.}

\item{units}{.}
}
\value{
S4 object
}
\description{
Title
}
