% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortClass.R
\docType{class}
\name{Cohort-class}
\alias{Cohort-class}
\alias{Cohort}
\title{Title}
\value{
T
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{start_date}}{T}

\item{\code{comorb_dx}}{T}

\item{\code{index_dx}}{T}

\item{\code{outcome_dx}}{T}

\item{\code{patient_lst}}{T}

\item{\code{end_date}}{T}
}}

