% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagnosisClass.R
\docType{class}
\name{Diagnosis-class}
\alias{Diagnosis-class}
\alias{Diagnosis}
\title{Title}
\value{
Obj
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{file_path}}{file path to diagnosis details}

\item{\code{code}}{character.}

\item{\code{type}}{character.}

\item{\code{descr_1}}{character.}

\item{\code{descr_2}}{character.}

\item{\code{descr_3}}{character.}

\item{\code{descr_4}}{character.}
}}

