% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientClass.R
\docType{class}
\name{Patient-class}
\alias{Patient-class}
\alias{Patient}
\title{Patient Class}
\value{
a new Patient object
}
\description{
Patient Class
}
\section{Fields}{

\describe{
\item{\code{id}}{character}

\item{\code{dob}}{Date}
}}

